Cleaned up FabricUtils
